requirement.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-wtf==1.1.1
flask-script==2.0.5
flask-login==0.6.2
Flask-Migrate==2.7.0
email_validator==2.0.0
mysql-connector-python==8.0.33